Move MetalShaderClassWrappingInterface to a separate file and generalise it
